moves stuff around
