+++ src/elfpy/agents/policies/single_short.py
@@ -1,7 +1,6 @@
 """User strategy that opens a single short and doesn't close until liquidation"""
-from elfpy.agent import Agent
-from elfpy.markets import Market
-from elfpy.types import MarketActionType
+from elfpy.agents.agent import Agent
+from elfpy.markets.hyperdrive import Market, MarketActionType
 
 # pylint: disable=duplicate-code
